Fraud Detection/services/drift-monitor/requirements.txt
faust-streaming
aiokafka<0.9.0
kafka-python-ng
